split ComputeQI2 interface, remove interfaces.qc
