+++ mriqc/utils/misc.py
@@ -428,3 +428,16 @@ def _flatten(in_dict, parent_key='', sep='_'):
     return dict(items)
 
 
+def _flatten_dict(indict):
+    out_qc = {}
+    for k, value in list(indict.items()):
+        if not isinstance(value, dict):
+            out_qc[k] = value
+        else:
+            for subk, subval in list(value.items()):
+                if not isinstance(subval, dict):
+                    out_qc['_'.join([k, subk])] = subval
+                else:
+                    for ssubk, ssubval in list(subval.items()):
+                        out_qc['_'.join([k, subk, ssubk])] = ssubval
+    return out_qc
