+++ mriqc/workflows/anatomical.py
@@ -7,7 +7,7 @@
 # @Date:   2016-01-05 11:24:05
 # @Email:  code@oscaresteban.es
 # @Last modified by:   oesteban
-# @Last Modified time: 2016-11-04 15:23:58
+# @Last Modified time: 2016-11-04 16:32:36
 """ A QC workflow for anatomical MRI """
 from __future__ import print_function, division, absolute_import, unicode_literals
 from builtins import zip, range
@@ -25,7 +25,7 @@ from niworkflows.anat.skullstrip import afni_wf as skullstrip_wf
 from niworkflows.anat.mni import RobustMNINormalization
 from mriqc.workflows.utils import fwhm_dict
 from mriqc.interfaces import (StructuralQC, ArtifactMask, ReadSidecarJSON,
-                              ConformImage)
+                              ConformImage, ComputeQI2)
 
 from mriqc.utils.misc import bids_getfile, bids_path, check_folder
 
@@ -145,9 +145,12 @@ def compute_iqms(settings, name='ComputeIQMs'):
     fwhm = pe.Node(afni.FWHMx(combine=True, detrend=True), name='smoothness')
     # fwhm.inputs.acf = True  # add when AFNI >= 16
 
+    # Mortamet's QI2
+    getqi2 = pe.Node(ComputeQI2(erodemsk=settings.get('testing', False)),
+                     name='ComputeQI2')
+
     # Compute python-coded measures
-    measures = pe.Node(StructuralQC(testing=settings.get('testing', False)),
-                       'measures')
+    measures = pe.Node(StructuralQC(), 'measures')
 
     # Project MNI segmentation to T1 space
     invt = pe.MapNode(ants.ApplyTransforms(
@@ -177,6 +180,8 @@ def compute_iqms(settings, name='ComputeIQMs'):
                                ('session_id', 'session_id'),
                                ('run_id', 'run_id'),
                                ('metadata', 'metadata')]),
+        (inputnode, getqi2, [('orig', 'in_file'),
+                             ('airmask', 'air_msk')]),
         (inputnode, measures, [('inu_corrected', 'in_noinu'),
                                ('in_inu', 'in_bias'),
                                ('orig', 'in_file'),
@@ -192,6 +197,7 @@ def compute_iqms(settings, name='ComputeIQMs'):
                            ('reverse_invert_flags', 'invert_transform_flags')]),
         (invt, measures, [('output_image', 'mni_tpms')]),
         (fwhm, datasink, [(('fwhm', fwhm_dict), 'fwhm')]),
+        (getqi2, datasink, [('qi2', 'qi2')]),
         (measures, datasink, [('summary', 'summary'),
                               ('spacing', 'spacing'),
                               ('size', 'size'),
@@ -203,12 +209,11 @@ def compute_iqms(settings, name='ComputeIQMs'):
                               ('fber', 'fber'),
                               ('efc', 'efc'),
                               ('qi1', 'qi1'),
-                              ('qi2', 'qi2'),
                               ('cjv', 'cjv'),
                               ('wm2max', 'wm2max'),
                               ('tpm_overlap', 'tpm_overlap')]),
         (out_name, datasink, [('out_file', 'out_file')]),
-        (measures, outputnode, [('out_noisefit', 'out_noisefit')]),
+        (getqi2, outputnode, [('out_file', 'out_noisefit')]),
         (datasink, outputnode, [('out_file', 'out_file')])
     ])
     return workflow
