+++ mriqc/workflows/functional.py
@@ -19,8 +19,7 @@ from nipype.interfaces import fsl
 from nipype.interfaces import afni
 
 from mriqc.workflows.utils import fmri_getidx, fwhm_dict, fd_jenkinson, thresh_image
-from mriqc.interfaces.qc import FunctionalQC
-from mriqc.interfaces.functional import Spikes
+from mriqc.interfaces.functional import FunctionalQC, Spikes
 from mriqc.utils.misc import bids_getfile, bids_path, check_folder, reorient_and_discard_non_steady
 
 DEFAULT_FD_RADIUS = 50.
